+++ neutpy.conf
@@ -16,7 +16,7 @@ verbose:1
 # The temperature coefficients calculation utilizes the Pathos multiprocessing package. If this value is not given,
 # neutpy runs on a single thread. If the number of CPU cores given is greater than the number of cores available,
 # neutpy will default to N-1, where N is the number of cores available
-cpu_cores:7
+cpu_cores:8
 
 # Core lines
 #
