+++ requirements.txt
@@ -1,19 +1,40 @@
+appdirs==1.4.3
 backports.functools-lru-cache==1.6.1
+CacheControl==0.12.6
+certifi==2019.11.28
+chardet==3.0.4
+colorama==0.4.3
+contextlib2==0.6.0
 cycler==0.10.0
 dill==0.3.2
+distlib==0.3.0
+distro==1.4.0
+html5lib==1.0.1
+idna==2.8
+ipaddr==2.2.0
 kiwisolver==1.1.0
+lockfile==0.12.2
 matplotlib==2.2.5
-multiprocess==0.70.10
+msgpack==0.6.2
+multiprocess==0.70.11.dev0
 numpy==1.16.6
+packaging==20.3
 pandas==0.24.2
-pathos==0.2.6
+pathos==0.2.7.dev0
+pep517==0.8.2
 pox==0.2.8
 ppft==1.6.6.2
+progress==1.5
 pyparsing==2.4.7
 python-dateutil==2.8.1
+pytoml==0.1.21
 pytz==2020.1
 PyYAML==5.3.1
+requests==2.22.0
+retrying==1.3.3
 scipy==1.2.3
 Shapely==1.7.0
 six==1.15.0
 subprocess32==3.5.4
+urllib3==1.25.8
+webencodings==0.5.1
